pyproject.toml
[project]
name = "pox-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiocache>=0.12.3",
    "aiofiles>=25.1.0",
    "aiohttp>=3.13.2",
    "aiomysql>=0.3.2",
    "aiosqlite>=0.21.0",
    "asyncpg>=0.31.0",
    "attr>=0.3.2",
    "deep-translator>=1.11.4",
    "deptry>=0.25.1",
    "discord-py>=2.6.4",
    "distro>=1.9.0",
    "edge-tts>=7.2.3",
    "gtts>=2.5.4",
    "i18nice>=0.16.0",
    "lmstudio>=1.5.0",
    "markovify>=0.9.4",
    "matplotlib>=3.10.7",
    "mcstatus>=12.0.6",
    "moviepy>=1.0.3",
    "ollama>=0.6.1",
    "openai>=2.29.0",
    "orjson>=3.11.8",
    "pillow>=12.0.0",
    "piper-tts>=1.3.0",
    "profanityfilter>=2.1.0",
    "proglog>=0.1.12",
    "prompt-toolkit>=3.0.52",
    "psutil>=7.2.2",
    "pymojang>=2.0.6",
    "pyqt6>=6.10.2",
    "pyqtgraph>=0.14.0",
    "python-dotenv>=1.2.1",
    "pytz>=2025.2",
    "pyyaml>=6.0.3",
    "qrcode[pil]>=8.2",
    "rich>=14.2.0",
    "roblox>=2.0.0",
    "thefuzz>=0.22.1",
    "websockets>=15.0.1",
]
